Fixed wrong types
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@5408 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/FORMAT/Base64.h
@@ -110,7 +110,7 @@ namespace OpenMS
 		@brief endianizes an unsigned long long. From BigEndian to LittleEndian and vice versa
 		@return returns the endianized number
 	*/
-		inline unsigned long long endianize64_(unsigned long long n);
+		inline UInt64 endianize64_(UInt64 n);
   };
 
 	inline UInt Base64::endianize32_(UInt n)
